+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- mTLS connections no longer disconnect when the certificate refresh command exits with a non-zero code
+
 ## 0.8.5 - 2026-02-03
 
 ### Added