+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.8.6 - 2026-03-05
+
 ### Changed
 
 - mTLS connections no longer disconnect when the certificate refresh command exits with a non-zero code